pyproject.toml
[tool.poetry]
name = "planperm"
version = "0.1.0"
description = ""
authors = ["GlobalAA <globala977@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tortoise-orm = "^0.21.6"
aiogram = "^3.12.0"
pydantic-settings = "^2.4.0"
celery = "^5.4.0"
redis = "^5.0.8"
rq = "^1.16.2"
rq-scheduler = "^0.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
